refactor: separate thread fetching and share split queries
